Fixing the undefined references. It is still necessary to check if there is an implementation for BCM_timer enable, disable and set methods. On PMU Init changes, it is important to remember Raspberry Pi3 work as ARMv7 architecture when running at 32 bits config. Removing the references to mailbox_eoi on ic_init, it compiled. TODO: fix the mailbox_eoi references on ic_init!

+++ include/machine/cortex/raspberry_pi3/raspberry_pi3_ic.h
@@ -39,8 +39,8 @@ public:
         mbox()->enable();
     }
 
-    static void enable(int i) {
-        assert((unsigned int)i <= INTS);
+    static void enable(unsigned int i) {
+        assert(i <= INTS);
         switch(i / 32){
         case 0 /* */:
         case 1 /* */:
@@ -53,8 +53,8 @@ public:
         mbox()->disable();
     }
 
-    static void disable(int i) {
-        assert((unsigned int)i <= INTS);
+    static void disable(unsigned int i) {
+        assert(i <= INTS);
         switch(i / 32){
         case 0 /* */:
         case 1 /* */:
